+++ CHANGELOG.md
@@ -26,3 +26,9 @@
 ## [1.2.3]
 
 - Fix scanner compatibility with asdf CLI manager
+- Remove deprecate install.sh usage
+
+## [1.2.4]
+
+- Update MCP Setup to use current XYGENI_TOKEN
+- Allow download scanner CLI from non-production Xygeni API Url (e.g. on-premise environments)